moved characteroverlay, overlay, toggle to lib folder
