+++ src/lib/components/character-overlay/index.module.scss
