+++ src/lib/components/index.ts
@@ -1,2 +1,5 @@
 export * from './unit-sheet';
 export * from './unit-sheet-sprite';
+export * from './character-overlay';
+export * from './overlay';
+export * from './toggle';
