+++ src/lib/components/overlay/index.tsx
@@ -1,4 +1,4 @@
 import styles from './index.module.scss';
-export default function OptionSelector({ onClick }: { onClick?: () => void }) {
+export function Overlay({ onClick }: { onClick?: () => void }) {
     return <div className={styles.overlay} onClick={onClick}></div>
 }
