+++ src/lib/components/toggle/index.module.scss
