+++ src/lib/components/toggle/index.tsx
@@ -1,6 +1,6 @@
 import styles from './index.module.scss';
 
-export default function Toggle({ active, onStateChange, ariaLabel }:
+export function Toggle({ active, onStateChange, ariaLabel }:
     { active: boolean, onStateChange: (isActive: boolean) => void, ariaLabel?: string }) {
     return <button className={`${styles.toggle} ${active ? styles.toggleActive : ''}`} 
         onClick={()=>onStateChange(!active)}
